Fix genesis configs (#269)
* Fix dev/local chain genesis

This commit fixes inconsistencies in the genesis configurations for dev and local chains, which also broke some
benchmarks.
Also, a new test is added to protect against unwanted future changes to the dev genesis used for benchmarking.

* Re-enable try-runtime for the CI

Now that we have a publicly available endpoint for the Volta testnet, we can integrate try-runtime checks with the CI.

* Add golden reference for dev chain genesis